+++ includes/class-disable-blog-activator.php
@@ -28,8 +28,11 @@ class Disable_Blog_Activator {
 	 * @since    0.4.3
 	 */
 	public static function activate() {
+
 		wp_cache_delete( 'comments-0', 'counts' );
 		delete_transient( 'wc_count_comments' );
+		flush_rewrite_rules( true );
+
 	}
 
 }