+++ includes/functions.php
@@ -39,11 +39,11 @@ function dwpb_post_types_with_feature( $feature ) {
 
 	/**
 	 * Filter the returned value.
-	 * 
+	 *
 	 * @since 0.4.0
-	 * 
+	 *
 	 * @param array|bool $post_types_with_feature an array of post types support this feature or false if none.
-	 * 
+	 *
 	 * @return array|bool
 	 */
 	return apply_filters( "dwpb_post_types_supporting_{$feature}", $post_types_with_feature );
@@ -66,7 +66,7 @@ function dwpb_post_types_with_feature( $feature ) {
  * @param string $taxonomy Required. The name of the feature to check against post type support.
  * @param array | string $args Optional. An array of key => value arguments to match against the post type objects. Default empty array.
  * @param string $output Optional. The type of output to return. Accepts post type 'names' or 'objects'. Default 'names'.
- * 
+ *
  * @return array | bool	A list of post type names or objects that have the taxonomy or false if nothing found.
  */
 function dwpb_post_types_with_tax( $taxonomy, $args = array(), $output = 'names' ) {