+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 ### Changelog
 
+##### 0.4.9
+- Extended method for disabling pingbacks to include internal pingbacks on posts and "update services."
+- Now fully disabling XMLRPC for posts and tags/categories. Tag/categories remain if another post type supports them.
+- Re-order admin menu to place "Pages" below "Dashboard," removing the divider after "dashboard."
+- Flush rewrite rules at activation and deactivation.
+- Filtering out `post` post types from all archives, previously it was just author archives and search results.
+- Removes post, category, and tag options from all menus. Tag/categories remain if another post type supports them.
+- Remove header feed urls, unless supported by another post type.
+- Expanded the approach used to disable REST API to completely turn off all public arguments on `post` post type and built-in taxonomies, unless another post type supports them.
+- Wordpress 5.5 support:
+    - Remove 'post' post type from XML sitemaps.
+    - Remove built-in taxonomies from XML sitemaps, if not being used by a custom post type.
+    - Fix sitemap redirect issues.
+- **Developers:** Some filters were removed and altered in this version:
+	- The `dwpb_redirect_feeds` filter now has (3) params, to match those in the `dwpb_disable_feed` filter: $bool, $post, $is_comment_feed.
+	- The `dwpb_disable_rest_api` filter was removed.
+	- The `dwpb_author_post_types` filter is now `dwpb_archive_post_types`, as the query modification now includes all pages passing `is_archive`.
+
 ##### 0.4.8.1
 - Do'h! Forgot to update the version number in the main plugin file. Bump.